refactor storytellingApi
