+++ web/src/services/api/storytellingApi/pages.ts
@@ -3,8 +3,6 @@ import { useCallback } from "react";
 
 import { MutationReturn } from "@reearth/services/api/types";
 import {
-  CreateStoryInput,
-  CreateStoryMutation,
   CreateStoryPageInput,
   CreateStoryPageMutation,
   CreateTeamPayload,
@@ -12,46 +10,24 @@ import {
   DeleteStoryPageMutation,
   MoveStoryPageInput,
   MoveStoryPageMutation,
-  MutationCreateStoryArgs,
   MutationCreateStoryPageArgs,
   MutationMoveStoryPageArgs,
   MutationRemoveStoryPageArgs,
 } from "@reearth/services/gql/__gen__/graphql";
 import {
-  CREATE_STORY,
   CREATE_STORY_PAGE,
   DELETE_STORY_PAGE,
   MOVE_STORY_PAGE,
 } from "@reearth/services/gql/queries/storytelling";
 import { useT } from "@reearth/services/i18n";
 
-import { useNotification } from "../state";
+import { useNotification } from "../../state";
 
 export type Team = CreateTeamPayload["team"];
 
-export default function useStorytellingAPI() {
-  const t = useT();
+export default () => {
   const [, setNotification] = useNotification();
-
-  const [createStoryMutation] = useMutation<CreateStoryMutation, MutationCreateStoryArgs>(
-    CREATE_STORY,
-  );
-  const useCreateStory = useCallback(
-    async (input: CreateStoryInput): Promise<MutationReturn<CreateStoryMutation>> => {
-      const { data, errors } = await createStoryMutation({
-        variables: {
-          input,
-        },
-      });
-      if (errors || !data?.createStory?.story?.id) {
-        setNotification({ type: "error", text: t("Failed to create story.") });
-
-        return { status: "error", errors };
-      }
-      return { data, status: "success" };
-    },
-    [createStoryMutation, setNotification, t],
-  );
+  const t = useT();
 
   const [createStoryPageMutation] = useMutation<
     CreateStoryPageMutation,
@@ -115,11 +91,9 @@ export default function useStorytellingAPI() {
     },
     [moveStoryPageMutation],
   );
-
   return {
-    useCreateStory,
     useCreateStoryPage,
     useDeleteStoryPage,
     useMoveStoryPage,
   };
-}
+};
